requirements.txt
yt-dlp
torch
ffmpeg-python
openai-whisper
pyautogui
pyperclip
streamlit